requirements.txt
python-dotenv
colorama
colored-traceback